test: align test file names with pixi convention and simplify scene renderers
Rename .x.test.ts files to match their source modules (createUboSyncSTD40,
PipelineSystem, GpuRenderTargetAdaptor) and fold GpuTextureSystem.viewKey into
GpuTextureSystem.test.ts under a new describe. Switch the color-read,
copy-depth-texture, and depth-read-only scenes to the array renderer form and
drop the unused canvas snapshots.
