+++ src/rendering/renderers/gpu/texture/__tests__/GpuTextureSystem.test.ts
@@ -1,3 +1,4 @@
+import { TextureSource } from '../../../shared/texture/sources/TextureSource';
 import { Texture } from '../../../shared/texture/Texture';
 import { describeLocalOnly, getWebGPURenderer } from '@test-utils';
 
@@ -57,5 +58,31 @@ describeLocalOnly('GpuTextureSystem', () =>
         const view2 = renderer.texture.getTextureRenderTargetView(texture, 0, 1); // Layer 1
 
         expect(view1).not.toBe(view2);
+    });
+});
+
+describeLocalOnly('GpuTextureSystem texture view cache key', () =>
+{
+    it('should return distinct views for descriptors differing only in mip/layer fields', async () =>
+    {
+        renderer = (await getWebGPURenderer()) as WebGPURenderer;
+
+        const source = new TextureSource({
+            width: 16, height: 16, mipLevelCount: 2, autoGenerateMipmaps: false,
+        });
+
+        const mip0 = renderer.texture.getTextureView(source, { baseMipLevel: 0, mipLevelCount: 1 });
+        const mip1 = renderer.texture.getTextureView(source, { baseMipLevel: 1, mipLevelCount: 1 });
+        const defaultView = renderer.texture.getTextureView(source);
+
+        // distinct subresources must not share a cached GPUTextureView — aliasing
+        // means silently sampling the wrong mip/layer with no error anywhere
+        expect(mip0).not.toBe(mip1);
+        expect(mip0).not.toBe(defaultView);
+
+        // identical descriptors must still hit the cache
+        const mip0Again = renderer.texture.getTextureView(source, { baseMipLevel: 0, mipLevelCount: 1 });
+
+        expect(mip0Again).toBe(mip0);
     });
 });
